Bill generations in hundredths of a coin, not whole ones (#51)
* Bill generations in hundredths of a coin, not whole ones

The ceil to a whole coin was larger than the price itself at the cheap
end. Z-Image costs $0.004, which is exactly 0.16 of a coin, and charging
1 coin for it billed 6.25x the margin the economics name. Recraft's
upscale was 10x over, background removal 5x. Those are the models a
plan's headline output count is quoted from, so the overcharge landed
precisely where the shop is compared against a competitor.

Nothing in storage ever needed the whole coin. The ledger has been BIGINT
micro-credits from the start and `model_prices` already holds exact
sub-coin rates — Flux 2 Flex is seeded as 2_800_000, not 3_000_000. The
whole coin was a display convention that had reached into the money path,
and `resolvePrice` was rounding away a precision the rows already had.

Billing now rounds up to MICRO_CREDITS_PER_BILLED_STEP, a hundredth of a
coin. Still upward: that is the only direction that cannot sell a job
below cost. Fine enough that no offered model rounds up by more than a
fraction of a percent, coarse enough that a quote is two decimals rather
than a float tail.

41 of the 81 benchmark rows get cheaper, 40 are unchanged — a model
already priced at a whole number of coins does not move, which is why
Kling 3.0 and the expensive video rows are untouched. Creator's headline
image estimate goes from 1,675 to 2,791. Plan margins are unchanged:
planMargin is MARGIN x price / (COIN_USD x coins) and contains no
per-model term.

Coin amounts stop being integers, so the quote, job, and wallet contracts
drop `.int()`. Plan grants stay whole (coinsToMicroCredits still refuses a
fraction) and timestamps are untouched. microCreditsToCoins now floors to
the billed step instead of the whole coin: a wallet holding 0.9 of a coin
can pay for five Z-Image runs, and showing that as zero would push a
customer to top up when they did not need to.

microCreditsFor rounds to a whole micro-credit before the step ceil.
(0.004 * 2) / 0.05 * 1e6 is 160000.00000000003 in doubles, and a bare
ceil on that noise bills 0.17 for a price that is exactly 0.16.

pricing.expected.json is re-frozen: 411 of its 738 combinations repriced.
That fixture exists to catch an unintended price change, so the header now
records that this one was deliberate.


* Update the two integration tests that asserted whole-coin billing

Both mirror unit tests already updated in the previous commit, and both
fail for the reason the change intends rather than for a defect.

pricingRepository: a row seeded at 2_400_000 micro-credits now charges
2.4 coins and stores 2_400_000, where it used to round to a whole 3. Added
a second case at 1_234_500 so the upward rounding is still pinned — a
price between steps must go up, never down, or the job sells below cost.

walletRepository: the test guards against converting each credit lot to
coins and adding, instead of summing micro-credits and converting once.
Three half-coin lots no longer discriminate, because half a coin is now a
whole number of billed steps and both routes give 1.5. Three lots of 5_000
do: each is below one step, so per-lot conversion floors all three to
nothing. Six lots together total 1.51, and the per-lot route reports 1.50.

Verified by running the resolver and microCreditsToCoins directly against
these values — that is everything the two tests exercise beyond the
database round-trip, which needs a container this environment has no
docker for.


---------

+++ packages/contracts/src/generation.ts
@@ -33,10 +33,14 @@ export const QuoteGenerationRequestSchema = z
  * `coins` is authoritative and always present — zero is a real answer, not a
  * missing one. `unlimited` is present only when the zero came from a grant
  * rather than from a free price, so the UI can say why and show what is left.
+ *
+ * Not a whole number: generations bill in hundredths of a coin, so a cheap
+ * model quotes 0.16 instead of rounding up to 1. See
+ * MICRO_CREDITS_PER_BILLED_STEP in @vgen/core.
  */
 export const GenerationQuoteSchema = z.object({
   id: z.uuid(),
-  coins: z.number().int().nonnegative(),
+  coins: z.number().nonnegative(),
   expiresAt: z.number().int().nonnegative(),
   unlimited: z
     .object({
@@ -115,8 +119,8 @@ export const GenerationJobSchema = z.object({
   status: GenerationJobStatusSchema,
   familyId: z.string().min(1),
   variantId: z.string().min(1),
-  /** What it was quoted at. Also what it was charged, once it succeeds. */
-  coins: z.number().int().nonnegative(),
+  /** What it was quoted at, in hundredths of a coin. Also what it was charged, once it succeeds. */
+  coins: z.number().nonnegative(),
   prompt: z.string(),
   createdAt: z.number().int().nonnegative(),
   updatedAt: z.number().int().nonnegative(),
